Pass `version` through all `_process_shard` calls
We need to pass the `_version` kwarg to `defer`, but we also need to pass the `version` arg so that subsequent re-defers of `_process_shard` can keep on re-deferring to the same version.

There's no environment var that tells us the current version, and even if there was, it wouldn't help with the case of "default" (i.e. "always run on the latest version). So we have to pass the version through all the calls.